+++ edu/slavik/interstripation/core/standard/VerticalInterstripation.py
@@ -10,14 +10,14 @@ class VerticalInterstripation:
     @staticmethod
     def restore(surface, stripe):
         """
-        Restore given vertical stripe data in given surface
+        Restore given vertical defect data in given surface
 
         Parameters
         ----------
         surface: Surface
             surface data
         stripe: Stripe
-            stripe data
+            defect data
         """
         startAt = stripe.startAt - 1
         endAt = stripe.endAt
