+++ edu/slavik/interstripation/defect/Region.py
