+++ edu/slavik/interstripation/defect/VerticalStripe.py
@@ -1,12 +1,12 @@
-from edu.slavik.interstripation.stripe.Stripe import Stripe
+from edu.slavik.interstripation.defect.Stripe import Stripe
 """
-Implementation of Stripe class for horizontal stripes, edges of which are parallel to abscissa axis 
+Implementation of Stripe class for vertical stripes, edges of which are parallel to ordinate axis 
 
 @author: oleksii.slavik 
 """
 
 
-class HorizontalStripe(Stripe):
+class VerticalStripe(Stripe):
 
     def __init__(self, startAt, endAt):
         """
@@ -15,16 +15,16 @@ class HorizontalStripe(Stripe):
         Parameters
         ----------
         startAt: int
-            left edge of stripe
+            top edge of defect
         endAt: int
-            right edge of stripe
+            bottom edge of defect
         """
         self.startAt = startAt
         self.endAt = endAt
 
     def removeStripe(self, surface):
         """
-        Remove current horizontal stripe from surface
+        Remove current vertical defect from surface
 
         Parameters
         ----------
@@ -34,12 +34,12 @@ class HorizontalStripe(Stripe):
         Return
         ------
         surface: Surface
-            surface without current horizontal stripe
+            surface without current vertical defect
         """
-        width = surface.matrix.shape[0]
+        height = surface.matrix.shape[1]
 
-        for i in range(self.startAt, self.endAt):
-            for j in range(width):
+        for i in range(height):
+            for j in range(self.startAt, self.endAt):
                 surface.matrix[i, j] = -1
 
         return surface
